mloggerai-py/pyproject.toml
[build-system]
requires = ["setuptools>=40.8.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mloggerai"  # nome unico su PyPI
version = "0.0.0.dev7"
description = "MLoggerAI: AI-powered Python traceback analyzer"
readme = { file = "README.md", content-type = "text/markdown" }
authors = [
  { name = "Mirko Perrone" },
]
requires-python = ">=3.10"
dependencies = [
    "openai>=1.109.1",
    "python-dotenv>=1.1.1",
]
license = "MIT"
license-files = ["LICENSE*"]

[project.urls]
Homepage = "https://github.com/perronemirko/mloggerai"
Issues = "https://github.com/perronemirko/mloggerai/issues"

[tool.setuptools.packages.find]
where = ["."]
